login and logout setup

+++ src/utils/cloudinary.js
@@ -1,3 +1,6 @@
+import dotenv from "dotenv";
+dotenv.config({ path: './.env' });
+
 import { v2 as cloudinary } from 'cloudinary'
 import fs from "fs"
 
@@ -9,22 +12,30 @@ cloudinary.config({
 });
 
 const uploadOnCloudinary = async (localFilePath) => {
+  // let safePath;
   try {
-       if(!localFilePath) return null
+       if(!localFilePath) return null;
+        // FIX 1: Convert Windows backslashes to forward slashes
+    // Cloudinary often fails with "C:\path\to\file"
+      //  const safePath = localFilePath.replace(/\\/g, "/");
        //upload the file on cloudinary
        const response = await cloudinary.uploader.upload(localFilePath, {
         resource_type: "auto"
        })
        //file has been uploaded successfully
-       console.log("file is uploaded on cloudinary", response.url);
+       //console.log("file is uploaded on cloudinary", response.url);
+       fs.unlinkSync(localFilePath)
        return response;
 
   } catch (error) {
-       fs.unlinkSync(localFilePath)  // remove the locally saved temporary file as the upload operation got failed
+       if (safePath) {
+         fs.unlinkSync(localFilePath);   // remove temporary file
+       }
+       console.log("CLOUDINARY UPLOAD ERROR:", error);
        return null;
   }
-}
+};
 
 
 
-export {uploadOnCloudinary}
+export {uploadOnCloudinary};
